+++ tilepy/include/PointingTools.py
@@ -1207,10 +1207,10 @@ def TransformRADec(vra,vdec):
             dec.append(coord.dec.deg)
     else:
         #print(vra,vdec)
-        ra = vra.astype(np.float)
-        dec = vdec.astype(np.float)
-        #np.float(vra)
-        #dec = np.float(vdec)
+        ra = vra.astype(float)
+        dec = vdec.astype(float)
+        #float(vra)
+        #dec = float(vdec)
     #print(ra,dec)
     coordinates = co.SkyCoord(ra, dec, frame='fk5', unit=(u.deg, u.deg))
     return coordinates
@@ -2485,10 +2485,10 @@ def LoadGalaxies_GladeCTASimu(tgalFile):
 
     name, dist, z, ra, dec, flag = np.genfromtxt(tgalFile, usecols=(0, 1, 2, 3, 4, 5), skip_header=3, unpack=True,
                                                  dtype='str')  # ra, dec in degrees
-    ra = ra.astype(np.float)
-    dec = dec.astype(np.float)
-    z = z.astype(np.float)
-    dist = dist.astype(np.float) / 1000  # change to Mpc!
+    ra = ra.astype(float)
+    dec = dec.astype(float)
+    z = z.astype(float)
+    dist = dist.astype(float) / 1000  # change to Mpc!
     tcat = Table([name, ra, dec, dist, z, flag], names=('Galaxy', 'RAJ2000', 'DEJ2000', 'Dist', 'z', 'flag'))
     # print(tcat)
     return tcat
@@ -2513,8 +2513,8 @@ def PointingFileReadCTA(pointingFile):
                                                                                  skip_header=1, unpack=True,
                                                                                  dtype='str')
     time = []
-    RA = RA.astype(np.float)
-    Dec = Dec.astype(np.float)
+    RA = RA.astype(float)
+    Dec = Dec.astype(float)
     for i in range(len(time1)):
         time.append((time1[i] + ' ' + time2[i]).split('"')[1])
     OutputTable = Table([time, RA, Dec, Observatory, ZenIni, ZenEnd, Duration],
@@ -2526,15 +2526,15 @@ def PointingFileReadCTA(pointingFile):
 def TableImportCTA_simple(inputFileName):
     run, MergerID, RA, Dec, distance, z, theta, ndet, SNR, A90, A50 = np.genfromtxt(inputFileName, usecols=(
     0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10), skip_header=3, unpack=True, dtype='str')
-    RA = RA.astype(np.float)
-    Dec = Dec.astype(np.float)
-    z = z.astype(np.float)
-    distance = distance.astype(np.float)
-    theta = theta.astype(np.float)
-    ndet = ndet.astype(np.float)
-    SNR = SNR.astype(np.float)
-    A90 = A90.astype(np.float)
-    A50 = A50.astype(np.float)
+    RA = RA.astype(float)
+    Dec = Dec.astype(float)
+    z = z.astype(float)
+    distance = distance.astype(float)
+    theta = theta.astype(float)
+    ndet = ndet.astype(float)
+    SNR = SNR.astype(float)
+    A90 = A90.astype(float)
+    A50 = A50.astype(float)
     OutputTable = Table([run, MergerID, RA, Dec, distance, z, theta, ndet, SNR, A90, A50], names=(
     'run', 'MergerID', 'RA', 'Dec', 'Distance', 'redshift', 'theta', 'ndet', 'SNR', 'A90', 'A50'))
     return OutputTable
@@ -2549,17 +2549,17 @@ def TableImportCTA(tgalFile):
                                                                                                       skip_header=1,
                                                                                                       unpack=True,
                                                                                                       dtype='str')
-    RA = RA.astype(np.float)
-    Dec = Dec.astype(np.float)
-    z = z.astype(np.float)
-    distance = distance.astype(np.float) / 1000  # to Mpc!
-    distMax = distMax.astype(np.float) / 1000  # to Mpc!
-    distMin = distMin.astype(np.float) / 1000  # to Mpc!
-    theta = theta.astype(np.float)
-    ndet = ndet.astype(np.float)
-    SNR = SNR.astype(np.float)
-    A90 = A90.astype(np.float)
-    A50 = A50.astype(np.float)
+    RA = RA.astype(float)
+    Dec = Dec.astype(float)
+    z = z.astype(float)
+    distance = distance.astype(float) / 1000  # to Mpc!
+    distMax = distMax.astype(float) / 1000  # to Mpc!
+    distMin = distMin.astype(float) / 1000  # to Mpc!
+    theta = theta.astype(float)
+    ndet = ndet.astype(float)
+    SNR = SNR.astype(float)
+    A90 = A90.astype(float)
+    A50 = A50.astype(float)
     OutputTable = Table([run, MergerID, RA, Dec, distance, distMin, distMax, z, theta, ndet, SNR, A90, A50], names=(
     'run', 'MergerID', 'RA', 'Dec', 'Distance', 'DistMin', 'DistMax', 'redshift', 'theta', 'ndet', 'SNR', 'A90', 'A50'))
     # print(OutputTable)
@@ -2569,15 +2569,15 @@ def TableImportCTA(tgalFile):
 def TableImportCTA_Glade(tgalFile):
     run, gal, MergerID, RA, Dec, distance, z, theta, ndet, SNR, A90, A50 = np.genfromtxt(tgalFile, usecols=(
     0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11), skip_header=2, unpack=True, dtype='str')
-    RA = RA.astype(np.float)
-    Dec = Dec.astype(np.float)
-    z = z.astype(np.float)
-    distance = distance.astype(np.float) / 1000  # to Mpc!
-    theta = theta.astype(np.float)
-    ndet = ndet.astype(np.float)
-    SNR = SNR.astype(np.float)
-    A90 = A90.astype(np.float)
-    A50 = A50.astype(np.float)
+    RA = RA.astype(float)
+    Dec = Dec.astype(float)
+    z = z.astype(float)
+    distance = distance.astype(float) / 1000  # to Mpc!
+    theta = theta.astype(float)
+    ndet = ndet.astype(float)
+    SNR = SNR.astype(float)
+    A90 = A90.astype(float)
+    A50 = A50.astype(float)
     OutputTable = Table([run, gal, MergerID, RA, Dec, distance, z, theta, ndet, SNR, A90, A50], names=(
     'run', 'Galaxy', 'MergerID', 'RA', 'Dec', 'Distance', 'redshift', 'theta', 'ndet', 'SNR', 'A90', 'A50'))
     # print(OutputTable)
@@ -2636,9 +2636,9 @@ def TableImportCTA_Obs(tobsFile):
 
 def TableImportCTA_LS(tgalFile):
     eventid, RA, Dec, distance = np.genfromtxt(tgalFile, usecols=(0, 3, 4, 8), skip_header=1, unpack=True, dtype='str')
-    RA = RA.astype(np.float)
-    Dec = Dec.astype(np.float)
-    distance = distance.astype(np.float) / 1000  # to Mpc!
+    RA = RA.astype(float)
+    Dec = Dec.astype(float)
+    distance = distance.astype(float) / 1000  # to Mpc!
     OutputTable = Table([eventid, RA, Dec, distance], names=('MergerID', 'RA', 'Dec', 'Distance'))
     return OutputTable
 
@@ -2771,7 +2771,7 @@ def ProduceSummaryFile(Source, SuggestedPointings, totalPoswindow, ID, obspar, t
     Pointings = SkyCoord(SuggestedPointingsC['RA[deg]'], SuggestedPointingsC['DEC[deg]'], frame='fk5',
                         unit=(u.deg, u.deg))
 
-    totalPGW = np.float('{:1.4f}'.format(np.float(sum(SuggestedPointingsC['PGW']))))
+    totalPGW = float('{:1.4f}'.format(float(sum(SuggestedPointingsC['PGW']))))
 
     # Check if the source is covered by the scheduled pointings
     Found, nP = IsSourceInside(Pointings, Source, obspar.FOV, obspar.ReducedNside)
